Revert React tree
