--- packages/rich-text/src/to-object-tree.js
@@ -1,63 +1,0 @@
-/**
- * Internal dependencies
- */
-import { toTree } from './to-tree';
-
-export function toObjectTree( options ) {
-	return toTree( {
-		...options,
-		createEmpty,
-		append,
-		getLastChild,
-		getParent,
-		isText,
-		getText,
-		remove,
-		appendText,
-	} );
-}
-
-function createEmpty() {
-	return {};
-}
-
-function getLastChild( { children } ) {
-	return children && children[ children.length - 1 ];
-}
-
-function append( parent, object ) {
-	if ( typeof object === 'string' ) {
-		object = { text: object };
-	}
-
-	object.parent = parent;
-	parent.children = parent.children || [];
-	parent.children.push( object );
-	return object;
-}
-
-function appendText( object, text ) {
-	object.text += text;
-}
-
-function getParent( { parent } ) {
-	return parent;
-}
-
-function isText( { text } ) {
-	return typeof text === 'string';
-}
-
-function getText( { text } ) {
-	return text;
-}
-
-function remove( object ) {
-	const index = object.parent.children.indexOf( object );
-
-	if ( index !== -1 ) {
-		object.parent.children.splice( index, 1 );
-	}
-
-	return object;
-}
